requirements.txt
streamlit==1.31.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.24.3
matplotlib==3.8.0
